fix(DTFS2-8076): address pr comments
